proplem 2
On your last dinner with a group of your friends, you and your friends decided to plan a Omrah trip during the month of Ramadan. Since you were feeling energetic that day, you decided to be in charge of planing the trip, which consists in:

Collecting from every member of the group the preferred period for the trip.
Check the hotel availability and prices during Ramadan. Upon checking several hotel, you made a list of hotels which have available rooms during the whole month along with the price for each night.
To schedule the trip, you have to respect several considerations:

The trip date should be within the preferred travel period of all group members.
The trip duration should be the longest possible given that it satisfies the first condition.
The start and end dates of the trip must be different.
You have to choose the cheapest hotel for your stay.
Your goal is to write a program to help you pick the dates of travel and the hotel where you should stay.

Input Format

The input consists of the description of a single problem instance. The description is given in several lines of text as follows.

The first line indicates the group size .
The following  lines contain the preferred period for each member of the group. The period is specified by the first and end days, which are given as integers between 1 and 30. The period is inclusive of start and end days.
The following line indicates the number of hotels .
The following  lines contain the prices for the hole month Ramadan at each hotel. Each line contains the prices of 30 nights for one hotel as integers separated by a space. For example, in the sample input below, there are 4 members in the group (the first member wants to go on the 2nd of Ramadan and return on the 7th). There are 5 available hotels with the prices per night indicated in each line.
Constraints

Output Format

The output consists of two lines. The first line contains either "success" if you successfully find a date for the trip, or "fail" if there is no agreement among the group members about the travel dates. If the first line is "success", then the second line should display the total hotel costs as an integer. If the first line is "fail", then the second line must be 0. For the sample input, the output is given below (the travel dates are 3 to 6 and the cheapest hotel is number 3. Note that you stay only 3 nights in the hotel: Ramadan 3rd, 4th and 5th).

Sample Input 0

4
2 7
3 7
1 6
3 9
5
35 34 31 37 32 37 39 30 32 35 37 36 30 33 38 36 32 35 36 35 33 39 31 37 31 34 40 30 39 38
20 50 29 39 36 24 33 29 46 30 49 34 35 23 20 44 37 34 39 33 37 45 48 24 43 44 37 49 45 34
20 11 23 14 22 25 12 22 25 25 23 16 17 18 13 26 12 22 22 20 16 17 27 21 25 17 22 24 15 25
30 36 31 39 35 31 38 41 59 55 57 56 48 56 36 43 50 40 41 44 33 38 58 50 41 45 40 52 57 30
40 36 34 25 23 33 39 46 37 35 27 42 23 26 27 47 32 43 49 44 46 31 26 36 28 47 37 47 24 43
Sample Output 0

success
59
Contest ends in 3 hours
Submissions: 0
Max Score: 20
Difficulty: Easy
Rate This Challenge: